Rules: Implement the rules processing engine
Note: Rules are currently implemented as a simplified python script that is
evaluated in a minimal and regulated namespace.  This is insecure and
sub-optimal and will eventually replaced with a real grammar.  However, the
rest of the concepts implemented in this patch should remain relavent even with
a new grammer.

To enable guest-specific reconfiguration, rules are processed once for each
monitored guest.  A namespace is constructed of three Entities: the Host data,
the Guest data, and an empty Output Entity.  Rules can access the value of a
statistic using <Entity>.<Statistic> syntax (ie. Host.total or
Guest.libvirt_maxmem).  The availability of statistics depends on which
Collectors are selected in the config file and rule-sets that access undefined
statistics will be discarded.  The Entity object provides rule-accessible
methods to access Monitor properties and calculate the average of a statistic.
Additionally, rules can work with local variables, local functions, condition
statements, and basic arithmetic.  Reconfiguration is triggerred by creating
Controller-defined variables in the Output Entity.  For example, a guest can be
ballooned to a new target size by setting Output.balloon_target to an integer
value.  After successful evaluation of the rule-set, the Entities are passed to
eah configured Controller where the requested configuration changes are
applied.

Signed-off-by: Adam Litke <agl@us.ibm.com>