Collectors: Collection exception reporting
If a Collector encounters an error during collection and is unable to return a
complete and correct set of statistics, an exception should be reported so that
the Monitor can decide how to handle it.  When a Monitor catches this
exception, it assumes that a transient error has occurred and the current row
of invalid statistics is discarded.  The Monitor marks itself not ready to
avoid further rule processing and Controller activity until a successful
collection occurs.

Signed-off-by: Adam Litke <agl@us.ibm.com>