Collectors: Handle temporary libvirt stats collection error
Until a guest initializes a virtio-balloon device driver with support for the
virtio memory statistics API, the GuestLibvirt Collector will not be able to
return a complete set of statistics.  This can occur during guest bootup before
the the driver has been loaded and also on guests with older balloon drivers.
Raise a CollectionError to avoid tripping up rule-sets that depend on the full
set of statistics.

Signed-off-by: Adam Litke <agl@us.ibm.com>