libvirt: Harden the libvirt interface
The libvirtInterface class exists to consolidate libvirt error handling in one
place so that it is not peppered throughout the rest of the source.  In order
to do this properly, calls to the libvirt API can only be made from within
libvirtInterface.  This patch replaces all remaining out of bounds API calls
with interface helpers.  Additionally, the reconnect logic is improved so that
the daemon can withstand libvirt restarts and crashes.  When libvirt goes down,
all GuestMonitors stop and the GuestManager continues to try connecting.  Once
the connection is restored, the GuestMonitors are respawned and processing
continues normally.

Signed-off-by: Adam Litke <agl@us.ibm.com>