Enable Monitor statistics plotting
To evaluate the effectiveness of the Memory Overcommit Manager, it is important
to be able to audit its behavior and the effect it has on the host and guests.
This patch introduces a mechanism that plots host and guest statistics as they
are collected.  The stats are written to a set of plot files that can be used
with gnuplot or other graphing software.  To enable this feature, create a
writable directory and set the config file variable 'plot-basedir' to that
directory.  A new directory will be created within plot-basedir for each
invocation of the program.

Signed-off-by: Adam Litke <agl@us.ibm.com>