Collector: Mechanism to report permanent Collector errors
If a collector is enabled in the mom config file, the user expects that
collector to run and it is likely that the installed rules depend on its data.
Therefore, if a collector encounters a premanent error that will prevent it
from working an error should be reported and the program terminated.  This
patch implements this error path and makes some changes to allow threads to be
cleaned up that have never been started.

Signed-off-by: Adam Litke <agl@us.ibm.com>