KSM: Add a KSM tuning algorithm to the example rules
Add a basic KSM tuning algorithm to the example rules.  If the host free memory
(including caches) drops below a certain threshold and there are enough KSM
shareable pages, turn on ksm and manage the number of pages to scan between
minimum and maximum parameters.  If the host has enough free memory or there is
not enough shareable memory, switch off KSM to reduce overhead.

Signed-off-by: Adam Litke <agl@us.ibm.com>