Predictably handle overlapping Collector data
In some cases it may be desireable for multiple collectors to return the same
set of statistics.  For example, memory stats can be collected through the
libvirt collector and also through the NetworkDaemon collector.  Depending on
the software versions of a particular guest, it may not be able to use the
libvirt collector.  To handle this case, we enable both collectors and allow
the data to be collcted by at least one of the collectors.  Conflicts are
solved by giving precedence to collectors listed first in the config file.

Signed-off-by: Adam Litke <agl@us.ibm.com>