Allow rules to create persistent variables
Sometimes rule processing can create data that, if maintained, could be used
for future rule processing cycles.  For example, the auto-balloon rules
calculate a balloon target and direct the balloon controller to adjust the
guests.  There is no way to determine if a guest is honoring the balloon
requests.  If we could store the balloon target along with the rest of the
guest data, the next time we evaluate rules we can see if the guest has
complied by comparing 'balloon_target' to 'libvirt_curmem'.  This information
could then be used to trigger a response.

Persistent variables are implemented as Entity methods GetVar() and PutVar().
Each time a Monitor is interrogated to produce an Entity, the Monitor adds any
defined variables to the Entity.  When Entities are cleaned up after rule
processing, the parent Monitor is updated with the set of variables.  In this
way, the Monitor remains the place for persistent storage of data and the
Entity is merely a transient representation of a Monitor at a given time.

Signed-off-by: Adam Litke <agl@us.ibm.com>