GuestMonitor: Use domain name instead of domain id
If the libvirt daemon is restarted, all of the current GuestMonitors are
recycled and recreated.  Unfortunately, the domain IDs will change causing the
Plotter to create new files rather than appending to the existing ones.  This
is decidedly inconvienent.  If we name our GuestMonitors using something
persistent (the name) then we avoid this problem.  As a side-effect, seeing the
domain name in the logs is more intuitive than seeing domain IDs.

This patch introduces a little more churn than one might expect.  This is
because the GuestMonitor initialization process must be significantly
reordered.  We must collect domain information before initializing the Monitor
and Thread parent classes so that our thread name can be put to good use.

Signed-off-by: Adam Litke <agl@us.ibm.com>