Revert "Monitor threads for starvation and deadlocks"
This reverts commit b77efadf9e6e2860527b3f62236a559e0b472984.

In hindsight, the whole MomThread thing was a bad idea.  It introduces way too
much complexity into the threading model for what is essentially a debugging
feature.  Something like this would be much better implemented as a simple,
periodic thread stack dumper that would not need to be integrated into the core
program logic.

Signed-off-by: Adam Litke <agl@us.ibm.com>