GuestManager: Finer-grained use of guests_sem
Libvirt operations can take a long time to complete, especially when the system
is under heavy load.  Therefore, it is not wise for the GuestManager to hold
the guests_sem semaphore while making libvirt calls.  To fix this, we make the
following changes:

 - When spawning new GuestMonitors, only take the lock when actually adding the
newly-spawned thread to the list.  Note that there is a small chance of
spawning a Monitor that is not needed.  In such cases, it will be terminated
again under the lock.

 - When waiting for GuestMonitors to exit, don't hold locks.  We pop one item
off of the list at a time and wait for it.

 - Merge all thread monitoring into check_threads() which eliminates the need
for reap_old_guests().  The actions done while holding the lock will not block.

Signed-off-by: Adam Litke <agl@us.ibm.com>