Debug: Thread stack dumping
Add a simple thread stack dumper to the code base.  This dumper works with no
external Python modules on python-2.5 or later.  On 2.4 you need the
threadframe module which you can get from here:

http://pypi.python.org/pypi/threadframe

There are no users of this code committed in the source code.  This is simply a
debugging aid.

Signed-off-by: Adam Litke <agl@us.ibm.com>