libvirt: Reduce the number of calls into libvirt
Under stress testing, I found that our single libvirt connection was becoming
saturated.  As a result, all threads that use libvirt were unable to make
progress.  The best way to relieve this pressure is to reduce our use of
libvirt.  As it turns out, there are a couple of opportunties to do that:

 - In the GuestManager, spawn_guest_monitors() and check_threads() both use a
guest list from libvirt.  Rather than asking for it twice, we can share it
between the two functions.  The potential side-effects of this are not serious.
For example, check_threads() won't see any newly-created threads until the next
interval.

 - The GuestMonitor check for domainIsRunning() is redundant.  The GuestManager
already knows this information so let it handle termination of the
GuestMonitor.

Signed-off-by: Adam Litke <agl@us.ibm.com>