Collector GuestNetworkDaemon: Update the protocol for increased reliability
Currently, the client (host) establishes a new connection to each server
(guest) at each collection interval.  This turns out to be a lot of overhead --
too much, in fact, for guests that are under heavy memory pressure.  Switch to
a protocol where the connection is kept open and only reconnected when
necessary.

Client        | Server
--------------+-----------
connect       |
              |     accept
request stats |
              | send stats
request stats |
              | send stats
             ...
        <disconnect>

Signed-off-by: Adam Litke <agl@us.ibm.com>