Collectors: Clean up Collector/Monitor error reporting interface
In general, Collectors should not log messages directly.  They should raise
either CollectionError or FatalError with an appropriate error message.  The
Monitor will catch those exceptions and decide how to handle them.  The current
exception to that rule is the GuestLibvirt Collector because it is permitted to
return a subset of data without error but we want to log that this is
happening.

This patch does two things: clean up the interface to remove any direct
Collector logging (except as explained above) and log all CollectionErrors when
we are in debug mode.

Signed-off-by: Adam Litke <agl@us.ibm.com>