GuestMonitor: The IP address is optional
Since discovering the IP address is only important when using the
GuestNetworkDaemon Collector, allow the GuestMonitor to start up without one.
A missing IP address will be properly detected and handled as a CollectionError
by the GuestNetworkDaemon Collector.  This patch allows MOM to work without an
IP address helper program in configurations that don't require the
GuestNetworkDaemon Collector.

Signed-off-by: Adam Litke <agl@us.ibm.com>