GuestQemuProc: Handle missing stat file when guest exits
If the qemu process of our guest has exited we will trigger an exception when
trying to read /proc/<pid>/stat because that file will no longer exist.  Handle
this case by raising a CollectionError.  The GuestManager will notice the guest
has gone and this GuestMonitor will be cleaned up soon anyway.

Signed-off-by: Adam Litke <agl@us.ibm.com>