Rules: Get rid of Output entity
In preparation for processing all guests in one rule evaluation pass, we need
to reorganize Controller triggers (like balloon_target) so that the controllers
will know which guest the request was intended for.  This can be done by
setting triggers directly in the Entity that will be affected.

Remove the Output entity and add a new Control() rules API function.  Calling
this function of an Entity will set the given trigger on that Entity.  For
example:

Guest.Control('balloon_target', foo)

instead of:
Output.SetVar('balloon_target', foo)

Signed-off-by: Adam Litke <agl@us.ibm.com>