GuestMonitor: Use wide ps format when looking up guest pid
We use ps output to look up a guests pid given its uuid.  Depending on the
system configuration, ps might truncate its output and the uuid portion of the
qemu command line will not appear.  This causes a grep match failure and
prevents us from getting the pid.  The solution is to use wide format ps output
to avoid the truncation.

Signed-off-by: Adam Litke <agl@us.ibm.com>