Integrate a policy language evaluator into MOM
From the beginning of this project, the goal has been to implement a policy
language framework.  For expediency, the initial policies were written in
Python and evaluated by the Python interpreter itself.  While this works in
practice, there are some problems:

 - Scope: Python-based policies are written as a subset of the Python language
   by convention only.  In order to provide a truly supportable policy language,
   the appropriate and allowed functionality should be explicit.  For example,
   file I/O and sub-process creation are possible in Python but beyond the scope
   of MOM policies.

 - Security: As an arbitrator of system resources, MOM needs to run with
   elevated privileges.  In such an environment, it is beyond foolish to execute
   untrusted Python code and this is precisely what with prototype policy engine
   does!  This must be changed.

MOM policies amount to configuration files.  They are sophisticated, rich
configuration files and they require an intelligent evaluation engine.  This
patch imports two things into the MOM project: the SPARK parser (version 0.6.1)
and some initial work on a parser contributed by Anthony Liguori.  We will not
make modifications to SPARK.  Over the remainder of this patch series, I will
be enhancing the MOM policy language with additional features and integrating
it into MOM.

Signed-off-by: Adam Litke <agl@us.ibm.com>
Signed-off-by: Anthony Liguori <aliguori@linux.vnet.ibm.com>