Parser: Support external functions
This patch creates a mechanism for Python library functions available from
within the policy.  Possible uses include: math operations (abs, sin), random
number generation, etc.  It's true that these can all be added to the Evaluator
class as 'c_' methods, but as the number of these functions grows, it may be
nice to separate these type of functions from core language constructs such as
if and let.

Signed-off-by: Adam Litke <agl@us.ibm.com>