PolicyEngine: Classes to support the new policy language within MOM
This patch adds a pair of new clases (PolicyEngine and Policy) that will
replace SystemController and Rules (respectively).  PolicyEngine is a MOM
thread that reads in a policy file at startup, creates a Policy object,
evaluates the policy regularly using the latest data about the host and guests,
and passes the results to the Controllers.  The Policy object encapsulates the
data structures created by the Parser and provides a high-level interface for
evaluating the policy with a given set of Host and Guest Entities.

Signed-off-by: Adam Litke <agl@us.ibm.com>