Parser: Test harness for MOM policy language
Any piece of code should have unit tests, but tests are especially important for our policy engine.  The tests serve to define the expected behavior of the language and will help us to avoid introducing regressions when working on this code.  I used the standard unittest framework that is included with Python.

Run the test by executing the following from the root of the source:
	python mom/Policy/TestParser.py

Signed-off-by: Adam Litke <agl@us.ibm.com>