RPC: Add new RPC calls to get and set the MOM policy
For the first really useful feature, this patch introduces a pair of calls to
get and set the current policy.  This could be useful (for example) if you want
to prioritize a set of guests at particular times of the day.  Most of this
patch is boilerplate RPC-type code but I also reorganize the policy loading
code to make it directly callable from the RPC server thread.

Signed-off-by: Adam Litke <agl@us.ibm.com>