RPC: Add getActiveGuests
Add a new RPC call to get the names of guests that are active.  An active guest
is one that is collecting and reporting statistics correctly according to the
configuration.  This information can be useful for identifying guests that
cannot be managed by MOM due to misconfiguration or other reasons.

Signed-off-by: Adam Litke <agl@us.ibm.com>