name-to-ip: Limit to GuestNetworkDaemon Collector
IP address resolution is only needed by the GuestNetworkDaemon Collector.  With
the introduction of Collector-specific mom.conf sections, it is possible to
isolate the use of name-to-ip to the one Collector that actually needs it.
Also, update the sample mom.conf files to ensure that name-to-ip-helper appears
in the correct config file section.

Signed-off-by: Adam Litke <agl@us.ibm.com>